+++ decidim-core/app/models/decidim/coauthorship.rb
@@ -8,6 +8,8 @@ module Decidim
 
     validates :coauthorable, presence: true
 
+    after_commit :author_is_follower, on: [:create]
+
     def identity
       user_group || author
     end
@@ -18,6 +20,13 @@ module Decidim
     # @returns The Organization for the Coauthorable
     def organization
       coauthorable&.organization
+    end
+
+    def author_is_follower
+      return unless author.is_a?(Decidim::User)
+      return unless coauthorable.is_a?(Decidim::Followable)
+
+      Decidim::Follow.find_or_create_by!(followable: coauthorable, user: author)
     end
   end
 end