+++ decidim-debates/lib/decidim/debates/engine.rb
@@ -36,7 +36,7 @@ module Decidim
           badge.levels = [1, 5, 10, 30, 50]
           badge.reset = lambda do |user|
             debates = Decidim::Comments::Comment.where(
-              decidim_author_id: user.id,
+              author: user,
               decidim_root_commentable_type: "Decidim::Debates::Debate"\
             )
             debates.pluck(:decidim_root_commentable_id).uniq.count