+++ decidim-proposals/app/controllers/decidim/proposals/proposals_controller.rb
@@ -203,7 +203,7 @@ module Decidim
       end
 
       def proposal_draft
-        Proposal.from_all_user_identities(current_user).not_hidden.where(component: current_component).find_by(published_at: nil)
+        Proposal.from_all_author_identities(current_user).not_hidden.where(component: current_component).find_by(published_at: nil)
       end
 
       def ensure_is_draft